Code for loading textures moved to another file Fixed managing default textures. Fixed loading LethalPainting mods Added dummy module if no textures loaded, but it enables saving data in Paiting item for compapability
