+++ CustomPainting/CustomPaintingModule.cs
@@ -1,198 +1,29 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.IO;
 using System.Linq;
-using System.Net;
-using BepInEx;
-using BepInEx.Configuration;
 using BepInEx.Logging;
-using CustomPainting.Configuration;
 using Unity.Netcode;
 using UnityEngine;
-using Logger = BepInEx.Logging.Logger;
 
 namespace CustomPainting;
 
-internal class CustomPaintingModule
+internal class CustomPaintingModule : CustomPaintingModuleBase
 {
     private readonly ManualLogSource _logger;
-    private readonly Config _config;
-    private const string DefaultTextureNameFormat = "default{0}.png";
-
-    private static readonly string[] DefaultTexturesUrls =
-    [
-        "https://i.imgur.com/ePiClDl.png",
-        "https://i.imgur.com/yZCdjxh.png",
-        "https://i.imgur.com/bIb8wXG.png",
-        "https://i.imgur.com/kl1QiwG.png",
-        "https://i.imgur.com/2kHSzn6.png",
-        "https://i.imgur.com/rvdZYBH.png",
-        "https://i.imgur.com/tgJinSI.png",
-        "https://i.imgur.com/ZullTxg.png"
-    ];
 
     private readonly string[] _paintingFiles;
     private System.Random? _randomForTextureSelection;
     private readonly Dictionary<int, int> _paintingValues = new();
 
-    internal bool HasTextures => _paintingFiles.Any();
-
-    internal CustomPaintingModule(ConfigFile config, string fullModName)
+    internal CustomPaintingModule(IEnumerable<string> paintingFiles, ManualLogSource logger)
     {
-        _logger = Logger.CreateLogSource(fullModName);
-        _config = new(config);
-
-        var defaultDirectory = PrepareDefaultDirectory();
-
-        PrepareWebTextures(defaultDirectory);
-
-        var directories = PrepareAdditionalTextureDirectories(defaultDirectory);
-
-        _paintingFiles = GetTextureFiles(directories).ToArray();
-
-        _logger.LogInfo("------- CustomPaintings loaded -------");
+        _paintingFiles = paintingFiles.ToArray();
+        _logger = logger;
     }
 
-    private IEnumerable<string> PrepareUrls()
-    {
-        var urls = new List<string>();
-        if (_config.DefaultPaintings.Value) 
-            urls.AddRange(DefaultTexturesUrls);
-
-        if (!string.IsNullOrEmpty(_config.CustomUrls.Value)) 
-            urls.AddRange(_config.CustomUrls.Value.Split(',').Select(v => v.Trim()).ToArray());
-
-        return urls;
-    }
-
-    private static string PrepareDefaultDirectory()
-    {
-        var defaultDirectory = Path.GetFullPath(Path.Combine(Paths.PluginPath, Config.DefaultDirectory));
-
-        if (!Directory.Exists(defaultDirectory))
-        {
-            Directory.CreateDirectory(defaultDirectory);
-        }
-        return defaultDirectory;
-    }
-
-    private IEnumerable<string> PrepareAdditionalTextureDirectories(string defaultDirectory)
-    {
-        // Obtain paintings from mods installed by mod manager
-        var directories = Directory.GetDirectories(Paths.PluginPath, Config.DefaultDirectory, SearchOption.AllDirectories)
-            .ToList();
-
-        // Add paintings made for LethalPainting mod
-        directories.AddRange(Directory.GetDirectories(Paths.PluginPath, "LethalPaintings", SearchOption.AllDirectories));
-
-        directories.AddRange(
-            _config.AdditionalDirectories.Value.Split(',')
-                .Select(d => d.Trim())
-                .Select(d => Path.IsPathRooted(d) ? d : Path.Combine(Paths.BepInExRootPath, d))
-        );
-
-        directories = directories
-            .Select(Path.GetFullPath)
-            .Where(d => !d.Equals(defaultDirectory, StringComparison.InvariantCultureIgnoreCase))
-            .Distinct()
-            .OrderBy(d => d)
-            .ToList();
-
-        directories.Insert(0, defaultDirectory);
-        return directories;
-    }
-
-    private void PrepareWebTextures(string directoryDownloadTo)
-    {
-        var shouldDownload = _config.ForceDownload.Value;
-        if (_config.DefaultPaintings.Value)
-            shouldDownload |= !CheckDefaultTextureExisting(directoryDownloadTo);
-        else
-            DeleteDefaultTextures(directoryDownloadTo);
-
-        var urls = PrepareUrls();
-        shouldDownload |= Directory.GetFiles(directoryDownloadTo).Length < urls.Count();
-
-        if (shouldDownload)
-            DownloadTextures(directoryDownloadTo, urls);
-    }
-
-    private static bool CheckDefaultTextureExisting(string directory)
-    {
-        for (var i = 0; i < DefaultTexturesUrls.Length; i++)
-        {
-            if (!File.Exists(Path.Combine(directory, string.Format(DefaultTextureNameFormat, i))))
-                return false;
-        }
-
-        return true;
-    }
-
-    private void DownloadTextures(string directory, IEnumerable<string> urls)
-    {
-        var i = 1;
-        foreach (var url in urls)
-        {
-            var request = (HttpWebRequest)WebRequest.Create(url);
-            var response = (HttpWebResponse)request.GetResponse();
-            if (response.StatusCode == HttpStatusCode.OK)
-            {
-                var fileName = _config.DefaultPaintings.Value
-                    ? i < DefaultTexturesUrls.Length
-                        ? $"{directory}/default{i}.png"
-                        : $"{directory}/custom{i - DefaultTexturesUrls.Length - 1}.png"
-                    : $"{directory}/custom{i}.png";
-
-                using var fileStream = new FileStream(fileName, FileMode.Create);
-                response.GetResponseStream()?.CopyTo(fileStream);
-            }
-            else
-            {
-                if (_config.DefaultPaintings.Value)
-                {
-                    if (i < 9)
-                    {
-                        _logger.LogWarning($"Error downloading default image {i}");
-                    }
-                    else
-                    {
-                        _logger.LogWarning($"Error downloading custom image {i - 8}");
-                    }
-                }
-                else
-                {
-                    _logger.LogWarning($"Error downloading custom image {i}");
-                }
-            }
-            i++;
-        }
-    }
-
-    private static void DeleteDefaultTextures(string directory)
-    {
-        for (var i = 1; i < 9; i++) 
-            File.Delete($"{directory}/default{i}.png");
-    }
-
-    private static IEnumerable<string> GetTextureFiles(IEnumerable<string> directories)
-    {
-        var files = new List<string>();
-        foreach (var directory in directories)
-        {
-            if (!Directory.Exists(directory)) continue;
-
-            var allPaintingPaths = Directory.GetFiles(directory);
-            files.AddRange(allPaintingPaths);
-        }
-
-        files.Sort();
-        return files;
-    }
-
-    internal void PatchPaintingItem(Item paintingItem)
+    internal override void PatchPaintingItemCore(Item paintingItem)
     {
         SetPaintingVariants(paintingItem);
-        paintingItem.saveItemVariable = true;
     }
 
     private void SetPaintingVariants(Item paintingItem)
@@ -220,7 +51,7 @@ internal class CustomPaintingModule
         _logger.LogInfo("Textures updated");
     }
 
-    internal void GenerateTextureIndexForPainting(GrabbableObject painting)
+    internal override void GenerateTextureIndexForPainting(GrabbableObject painting)
     {
         if (!StartOfRound.Instance.IsServer && StartOfRound.Instance.shipHasLanded)
         {
@@ -258,7 +89,7 @@ internal class CustomPaintingModule
         }
     }
 
-    internal void ApplyTexturesForPaintingsInSpawnedScrap(NetworkObjectReference[] spawnedScrap)
+    internal override void ApplyTexturesForPaintingsInSpawnedScrap(NetworkObjectReference[] spawnedScrap)
     {
         foreach (var networkObjectReference in spawnedScrap)
         {
@@ -278,14 +109,14 @@ internal class CustomPaintingModule
         }
     }
 
-    internal void SetSeedForTextureIndexGenerator(int roundSeed)
+    internal override void SetSeedForTextureIndexGenerator(int roundSeed)
     {
         var seed = roundSeed + 300;
         _logger.LogDebug($"Used seed {seed}");
         _randomForTextureSelection = new(seed);
     }
 
-    internal int GetPaintingFileIndex(GrabbableObject painting)
+    internal override int GetPaintingFileIndex(GrabbableObject painting)
     {
         if (_paintingValues.TryGetValue(painting.GetInstanceID(), out var fileIndex))
         {
@@ -297,7 +128,7 @@ internal class CustomPaintingModule
         return default;
     }
 
-    internal void LoadPaintingTextureByIndex(GrabbableObject painting, int fileIndex)
+    internal override void LoadPaintingTextureByIndex(GrabbableObject painting, int fileIndex)
     {
         if (fileIndex < 0)
         {
@@ -318,7 +149,7 @@ internal class CustomPaintingModule
         _logger.LogInfo($"Loaded fileIndex ({fileIndex}) for object {painting.gameObject} ({painting.GetInstanceID()})");
     }
 
-    internal void ClearSelectedTextures()
+    internal override void ClearSelectedTextures()
     {
         _paintingValues.Clear();
         _randomForTextureSelection = null;
